Lelu vehicle mapping

+++ src/main/java/fi/vaylavirasto/sillari/model/VehicleModel.java
@@ -3,8 +3,8 @@ import lombok.Data;
 
 @Data
 public class VehicleModel {
-    private long id;
-    private long permitId;
+    private Integer id;
+    private Integer permitId;
     private String type;
     private String identifier;
 }
